Made some UI improvements
Added app icon and made some cuatomizations in dark and light theme. Both themes now look much better than previous.
Only task is to welcome image for the app ( On the signup screen as well as to be shown each time when user enters the app ).
Also don't forget to write terms and conditions for the app